refactor(scim): extract group-flow helpers from groups.go (#149 slice 3)
Third slice. Lifts the per-RPC support functions out of groups.go
into a sibling groups_helpers.go file:

- reconcileGroupMembers (member-set diff + add/remove events)
- buildGroupResource (group + members \xe2\x86\x92 SCIMGroup, with
  orphan-restoration fallback)
- restoreOrphanedGroup (re-creates the user_group + remaps when
  the SCIM mapping outlives its underlying group)
- extractMembers (flexible value-shape decoder)
- extractUserIDFromMemberFilter (SCIM filter-path parser)

groups.go: 1006 \xe2\x86\x92 780 LOC (-226). Pure file-shuffling \xe2\x80\x94
zero behaviour change.

groups.go now contains only the per-RPC HTTP handlers (list /
listFiltered / get / create / replace / patch + 3 patch dispatchers /
delete).

All SCIM v2 conformance tests still pass (209s integration suite).

Refs manchtools/power-manage-server#149 (audit F009).
